+++ tests/blocks/test_aionotify.py
@@ -5,7 +5,8 @@ import pytest
 from asynctest import Mock
 from helpers import misc, task
 
-from i3pyblocks import types, utils
+from i3pyblocks import types
+from i3pyblocks._internal import utils
 from i3pyblocks.blocks import aionotify as m_aionotify
