+++ tests/blocks/test_subprocess.py
@@ -2,7 +2,8 @@ import pytest
 from asynctest import Mock
 from helpers import misc
 
-from i3pyblocks import types, utils
+from i3pyblocks import types
+from i3pyblocks._internal import utils
 from i3pyblocks.blocks import subprocess as m_sub
